--- a/Sources/LeafKit/LeafRenderer.swift
+++ b/Sources/LeafKit/LeafRenderer.swift
@@ -1,4 +1,33 @@
 import NIOConcurrencyHelpers
+import Foundation
+
+public enum LeafError: LocalizedError {
+    case unknownError(String)
+    case unsupportedFeature(String)
+    case cachingDisabled
+    case keyExists(String)
+    case noValueForKey(String)
+    case unresolvedAST(String)
+    case noTemplateExists(String)
+    case cyclicalReference(String)
+
+    var errorDescription: String {
+        switch self {
+            case .unknownError(let message): return message
+            case .unsupportedFeature(let feature): return "\(feature) is not implemented"
+            case .cachingDisabled: return "Caching is globablly disabled"
+            case .keyExists(let key): return "Existing entry \(key): use insert with replace=true to overrride"
+            case .noValueForKey(let key): return "No cache entry exists for \(key)"
+            case .unresolvedAST(let key): return "Flat AST expected; \(key) has unresolved dependencies"
+            case .noTemplateExists(let key): return "No template found named \(key)"
+            case .cyclicalReference(let key): return "\(key) was referenced causing a cyclical loop"
+        }
+    }
+
+    init(_ message: String) {
+        self = .unknownError(message)
+    }
+}
 
 public var defaultTags: [String: LeafTag] = [
     "lowercased": Lowercased(),
@@ -6,7 +35,7 @@ public var defaultTags: [String: LeafTag] = [
 
 public struct LeafConfiguration {
     public var rootDirectory: String
-    
+
     public init(rootDirectory: String) {
         self.rootDirectory = rootDirectory
     }
@@ -14,42 +43,42 @@ public struct LeafConfiguration {
 
 public protocol LeafCache {
     func insert(
-        _ document: ResolvedDocument,
+        _ document: LeafAST,
         on loop: EventLoop
-    ) -> EventLoopFuture<ResolvedDocument>
+    ) -> EventLoopFuture<LeafAST>
     func load(
         documentName: String,
         on loop: EventLoop
-    ) -> EventLoopFuture<ResolvedDocument?>
+    ) -> EventLoopFuture<LeafAST?>
     var isEnabled : Bool { get set }
 }
 
 public final class DefaultLeafCache: LeafCache {
     let lock: Lock
-    var cache: [String: ResolvedDocument]
+    var cache: [String: LeafAST]
     public var isEnabled: Bool = true
-    
+
     public init() {
         self.lock = .init()
         self.cache = [:]
     }
-    
+
     public func insert(
-        _ document: ResolvedDocument,
+        _ document: LeafAST,
         on loop: EventLoop
-    ) -> EventLoopFuture<ResolvedDocument> {
+    ) -> EventLoopFuture<LeafAST> {
         self.lock.lock()
         defer { self.lock.unlock() }
-       if isEnabled {
+        if isEnabled {
             self.cache[document.name] = document
         }
         return loop.makeSucceededFuture(document)
     }
-    
+
     public func load(
         documentName: String,
         on loop: EventLoop
-    ) -> EventLoopFuture<ResolvedDocument?> {
+    ) -> EventLoopFuture<LeafAST?> {
         self.lock.lock()
         defer { self.lock.unlock() }
         return loop.makeSucceededFuture(self.cache[documentName])
@@ -121,7 +150,7 @@ public struct NIOLeafFiles: LeafFiles {
     public init(fileio: NonBlockingFileIO) {
         self.fileio = fileio
     }
-    
+
     public func file(path: String, on eventLoop: EventLoop) -> EventLoopFuture<ByteBuffer> {
         let openFile = self.fileio.openFile(path: path, eventLoop: eventLoop)
         return openFile.flatMapErrorThrowing { error in
@@ -144,7 +173,7 @@ public final class LeafRenderer {
     public let files: LeafFiles
     public let eventLoop: EventLoop
     public let userInfo: [AnyHashable: Any]
-    
+
     public init(
         configuration: LeafConfiguration,
         tags: [String: LeafTag] = defaultTags,
@@ -160,14 +189,15 @@ public final class LeafRenderer {
         self.eventLoop = eventLoop
         self.userInfo = userInfo
     }
-    
+
     public func render(path: String, context: [String: LeafData]) -> EventLoopFuture<ByteBuffer> {
-        let expanded = expand(path: path)
-        let document = fetch(path: expanded)
+        let document = fetchFlat(template: path)
         return document.flatMapThrowing { try self.render($0, context: context) }
     }
-    
-    func render(_ doc: ResolvedDocument, context: [String: LeafData]) throws -> ByteBuffer {
+
+    func render(_ doc: LeafAST, context: [String: LeafData]) throws -> ByteBuffer {
+        guard doc.flat == true else { throw LeafError.unresolvedAST(doc.name) }
+
         var serializer = LeafSerializer(
             ast: doc.ast,
             context: context,
@@ -176,65 +206,89 @@ public final class LeafRenderer {
         )
         return try serializer.serialize()
     }
-    
-    private func expand(path: String) -> String {
-        var path = path
+
+    private func expand(template: String) -> String {
+        var path = template
         // ignore files that already have a type
         if path.split(separator: ".").count < 2, !path.hasSuffix(".leaf") {
             path += ".leaf"
         }
-        
+
         if !path.hasPrefix("/") {
             path = self.configuration.rootDirectory.trailSlash + path
         }
         return path
     }
-    
-    private func fetch(path: String) -> EventLoopFuture<ResolvedDocument> {
-        let expanded = expand(path: path)
-        return cache.load(documentName: expanded, on: eventLoop).flatMap { cached in
-            guard let cached = cached else { return self.read(file: expanded) }
-            return self.eventLoop.makeSucceededFuture(cached)
+
+    private func fetchFlat(template: String) -> EventLoopFuture<LeafAST> {
+        return self.fetch(template: template).flatMapThrowing { ast in
+            guard let ast = ast else { throw LeafError.noTemplateExists(template) }
+            guard ast.flat else { throw LeafError.unresolvedAST(template) }
+            return ast
         }
     }
-    
-    private func read(file: String) -> EventLoopFuture<ResolvedDocument> {
-        let raw = readBytes(file: file)
-        
-        let syntax = raw.flatMapThrowing { raw -> [Syntax] in
-            var raw = raw
-            guard let template = raw.readString(length: raw.readableBytes) else { return [] }
-            var lexer = LeafLexer(name: file, template: template)
-            let tokens = try lexer.lex()
-            var parser = LeafParser(name: file, tokens: tokens)
-            return try parser.parse()
-        }
-        
-        return syntax.flatMap { syntax in
-            let dependencies = self.readDependencies(syntax.dependencies)
-            let resolved = dependencies.flatMapThrowing { dependencies -> ResolvedDocument in
-                let unresolved = UnresolvedDocument(name: file, raw: syntax)
-                let resolver = ExtendResolver(document: unresolved, dependencies: dependencies)
-                return try resolver.resolve(rootDirectory: self.configuration.rootDirectory)
-            }
-            
-            return resolved.flatMap { resolved in self.cache.insert(resolved, on: self.eventLoop) }
-        }
-    }
-    
-    private func readDependencies(_ dependencies: [String]) -> EventLoopFuture<[ResolvedDocument]> {
-        let fetchRequests = dependencies.map(self.fetch)
-        let results = EventLoopFuture.whenAllComplete(fetchRequests, on: self.eventLoop)
-        return results.flatMapThrowing { results in
-            return try results.map { result -> ResolvedDocument in
-                switch result {
-                case .success(let ob): return ob
-                case .failure(let e): throw e
+
+    private func fetch(template: String, chain: Set<String> = .init()) -> EventLoopFuture<LeafAST?> {
+        return cache.load(documentName: template, on: eventLoop).flatMap { cached in
+            guard let cached = cached else {
+                return self.read(name: template).flatMap { ast in
+                    guard let ast = ast else { return self.eventLoop.makeSucceededFuture(nil) }
+                    return self.resolve(ast: ast, chain: chain, forceCaching: true).map {$0}
                 }
             }
+            guard cached.flat == false else { return self.eventLoop.makeSucceededFuture(cached) }
+            return self.resolve(ast: cached, chain: chain).map {$0}
         }
     }
-    
+
+    // resolve is only guaranteed to try to resolve an AST to flatness, not to succeed
+    private func resolve(ast: LeafAST, chain: Set<String>, forceCaching: Bool = false) -> EventLoopFuture<LeafAST> {
+        guard ast.flat == false || forceCaching == true else { return self.eventLoop.makeSucceededFuture(ast) }
+
+        var chain = chain
+        _ = chain.insert(ast.name)
+        guard chain.intersection(ast.unresolvedRefs).count == 0 else {
+            return self.eventLoop.makeFailedFuture(LeafError.cyclicalReference(ast.name))
+        }
+
+        let fetchRequests = ast.unresolvedRefs.map { self.fetch(template: $0, chain: chain) }
+        let results = EventLoopFuture.whenAllComplete(fetchRequests, on: self.eventLoop)
+ 
+        return results.flatMapThrowing { results -> LeafAST in
+            let results = results
+            var externals: [String: LeafAST] = [:]
+            for result in results {
+                // skip any unresolvable references
+                guard let external = try result.get() else { continue }
+                externals[external.name] = external
+            }
+            // return new inlined AST that may potentially still have references that can't be resolved
+            return LeafAST(from: ast, referencing: externals)
+        }.flatMap { resolved in
+            guard forceCaching else { return self.eventLoop.makeSucceededFuture(resolved) }
+            return self.cache.insert(resolved, on: self.eventLoop)
+        }
+    }
+
+    private func read(name: String) -> EventLoopFuture<LeafAST?> {
+        let path = expand(template: name)
+        let raw = readBytes(file: path)
+
+        return raw.flatMapThrowing { raw -> LeafAST? in
+            var raw = raw
+
+            // MARK: Should this actually throw an error if readString can't read readableBytes?
+            let template = raw.readString(length: raw.readableBytes) ?? ""
+            var lexer = LeafLexer(name: name, template: template)
+            let tokens = try lexer.lex()
+            var parser = LeafParser(name: name, tokens: tokens)
+            let ast = try parser.parse()
+            return LeafAST(name: name, ast: ast)
+        }
+    }
+
+//    private func readDependencies... *obviated by new fetchFlat/fetch functions*
+
     private func readBytes(file: String) -> EventLoopFuture<ByteBuffer> {
         self.files.file(path: file, on: self.eventLoop)
     }
@@ -244,7 +298,7 @@ extension Array where Element == Syntax {
     var dependencies: [String] {
         return extensions.map { $0.key }
     }
-    
+
     private var extensions: [Syntax.Extend] {
         return compactMap {
             switch $0 {