This PR substantially refactors LeafKit to enhance the functionality, capabilities, stability and performance of the template processing chain.

Significant changes:

### **LeafAST**
A new struct that replaces both Resolved and UnresolvedDocuments. It contains both the original raw Parsed AST (equivalent to UnresolvedDocument), and if resolvable, a flat AST (equivalent to ResolvedDocument), and maintains state of flatness, original external references, and any current unresolved external references (if applicable)
- this allows caching a template to _maintain its base state_ so that in the future, updating a template can cascade through the reference tree without reloading files.
- typealiased to ResolvedDocument to prevent breaking possible 3rd party implementations of LeafCache, as there's 1:1 mapping between LeafAST.name & LeafAST.ast (when flat) - **POTENTIALLY BREAKING** this assumes an external LeafCache implementation does not attempt to modify cached documents itself.
- initializers for both creating from a raw [Syntax] AST, or an existing LeafAST referencing a dictionary of provided external LeafASTs.
- the referencing initializer is smarter than the old ExtensionResolver, and will compress .raw syntax to make the resolved AST smaller (though in practice on large documents this may be a minor improvement) - resolves #36 

### **LeafError**
An Error class to allow greater specificity of error handling during the processing pipeline

### **LeafCache**
- Adds a .remove() method to allow purging an entry, per my previous PR - resolves #42  
- Adds an .insert(... replace:) method to allow control over insertion behavior when a cached template exists **POTENTIALLY BREAKING** as there is no DEFINED expectation of behavior currently - currently inserting a document for a key that already exists would overwrite in DefaultLeafCache, but this probably isn't an intended action and should be clarified. A default implementation assumes replacing isn't allowed.
- Adds an entryCount method to return the number of cache entries (this was for my own testing convenience but couldn't hurt

### **DefaultLeafCache**
- Enhanced error checking

### **LeafRenderer**
Subsumes functionality from ExtendResolver - 
- the render-fetch-serialize chain is entirely handled inside LeafRenderer and does not use external functions or classes.
- Now protects against cyclical loops (ie, Template A extends B, but Template B extends A)
- Removes a duplicate call to expand(path) for getting a template file's full path **(this alone provides a serious bump in speed by itself to current master branch)**
- Normalizes caching behavior - previously documents were cached with their full path as the key/name. Now LeafASTs use only their referencing name - eg a cached document is in the cache at "a" instead of "/full/path/to/templates/a.leaf" which should contribute to some performance boost regardless of cache implementation.

### **expand(path) function**
Modified to be a bit less aggressive - previously would not try to add ".leaf" if a period was present anywhere in the template name, assuming that it represented an alternate file - eg using a template name "mytemplate.html" would not be changed to "mytemplate.html.leaf". Still maintains this behavior but isolates the period split to the last section if the path is multi-directory, so allows for storing templates with reference to domain so that "mydomain.com/mytemplate" will appropriately expand to "/root/view/path/mydomain.com/mytemplate.leaf", if one wants to have templates organized by domain name.

### **ExtendResolver**
Removed, as its functionality is replaced by internal functions of LeafRenderer and LeafAST
- LeafRenderer will inspect a LeafAST for unresolved refs, attempt to fetch them, and then submit to a LeafAST init to resolve.
- LeafAST extension resolving .init function takes a base LeafAST and an array of provided referenced ASTs, and inline to the new AST.

### **DocumentLoader & FileAccessProtocol**
Removed - was noted as being for testing, and relevant tests were updated to obviate it

### **Tests**
- General cleanup of test cases that previously used a removed function
- Linear speed test for the cache generates X number of raw templates, and then iterates Y times accessing it against an MTELG to test basic insert/load speed
- Random speed test for the cache generates X,Y,Z number of raw templates that reference each other in layers (X extending 2 random Y templates, and each Y template extending a flat Z template), and then iterates A times accessing it against an MTELG to test insert/load speed and extension resolving.

### **PERFORMANCE NOTES**
- Linear test - 10 flat templates, 1 million render calls against them
- Random test - 50 flat templates (layer 3), 200 templates referencing one of the flat templates (layer 2), 1,000 templates referencing two random layer 2 templates (layer 1). 1 million render calls against a random one of the 1,250 total templates.
- Each test was run 50 times on an 4GHz i7 Retina iMac
- Branches - master (current), master* (current with duplicate expand call removed), LeafAST

#### Linear
Branch | Min | Avg | Max | Avg Baseline | CPU Time | CPU Baseline
--- | --- | --- | --- | --- | --- | ---
master | 9.21s | 9.76s | 10.57s | 100% | 3m 5s | 100%
master* | 7.94s | 8.21s | 8.62s | 84.2% | 2m 39s | 86.1%
LeafAST | 6.63s | 7.16s | 7.48s | **71.9%** | 2m 12s | **71.3%**

#### Random
Branch | Min | Avg | Max | Avg Baseline | CPU Time | CPU Baseline
--- | --- | --- | --- | --- | --- | ---
master | 12.15s | 12.58s | 13.1s | 100% | 3m 35s | 100%
master* | 8.97s | 9.49s | 9.87s | **75.4%** | 3m 32s | 98.3%
LeafAST | 9.23s | 9.61s | 10.2s | **76.8%** | 2m 45s | **76.8%**

#### Summary
This branch is doing slightly more work during resolution as it has to double check the AST at points instead of assuming a provided AST is flat, but generally speaking, it's substantially faster than either master version at any point in the Linear test. While the removal of the duplicate expand() call in master helps in the Random test to achieve the same speed overall, this branch is doing substantially less work overall, and ~~this points at the mutex in DefaultLeafCache being the bottleneck at this point~~ subsequent tests point at LeafCache always wrapping cache results in Futures as being a bottleneck.