@tanner0101 2020-04-24
These changes are now available in [1.0.0-rc.1.5](https://github.com/vapor/leaf-kit/releases/tag/1.0.0-rc.1.5)